edit views and add to API-doc

+++ blog/views.py
@@ -1,5 +1,5 @@
 from flask import jsonify
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_identity
 from sqlalchemy.exc import IntegrityError
 
 from extentions import db
@@ -17,7 +17,8 @@ def list_posts():
     for p in posts:
         tmp = dict()
         tmp['id'] = p.id
-        tmp = {'slug': p.slug, 'title': p.title, 'summary': p.summary, 'content': p.content, 'category': p.category}
+        slug = p.slug.replace(' ', '-')
+        tmp = {'slug': slug, 'title': p.title, 'summary': p.summary, 'content': p.content, 'category': p.category}
         result.update({f'{p.id}': tmp})
     return jsonify(result), 200
 
@@ -25,7 +26,9 @@ def list_posts():
 @blog.route('/<string:slug>', methods=['GET'])
 def single_post(slug):
     post = Post.query.filter(Post.slug == slug).first()
-    result = {'slug': post.slug, 'title': post.title, 'summary': post.summary, 'content': post.content,
+    if not post:
+        return {'error': 'post not exist!'}, 404
+    result = {'post_id':post.id, 'slug': post.slug, 'title': post.title, 'summary': post.summary, 'content': post.content,
               'category': post.category}
     return jsonify(result), 200
 
@@ -33,6 +36,7 @@ def single_post(slug):
 @blog.route('/posts/new/', methods=['POST'])
 @jwt_required
 def create_post():
+    ID = User.query.filter_by(Email=get_jwt_identity()).first().id
     form = PostForm()
     if not form.validate_on_submit():
         return {'error': form.errors}, 400
@@ -42,6 +46,7 @@ def create_post():
     new_post.slug = form.slug.data
     new_post.summary = form.summary.data
     new_post.category = form.category.data
+    new_post.authorId = ID
     db.session.add(new_post)
     db.session.commit()
     return {'msg': 'Post created successfully'}, 201
@@ -50,6 +55,9 @@ def create_post():
 @blog.route('/posts/delete/<int:post_id>/', methods=['DELETE'])
 @jwt_required
 def delete_post(post_id):
+    ID = User.query.filter_by(Email=get_jwt_identity()).first().id
+    if not ID == Post.query.get(post_id).authorId:
+        return {'error': 'access denied!'}, 403
     post = Post.query.get_or_404(post_id)
     db.session.delete(post)
     db.session.commit()
